Sync query algorithm with backend
We used to pass a single timestamp and expect all
entries since this time, up to the present.

Now, we pass two timestamps, the first one is the
start if the interval, the second one is the end.
And we expect all entries in this interval (inclusive).

The start if the interval is always the latest entry
we have. The end of the interval is the current time.

Both times are represented as UNIX timestamps.