Do not let the client cache future requests
Imagine we have a client whose clock is in the future. Then when it
makes a request in which the end timestamp is in the future and we
cache it, the next client who makes a request for the same timestamp
will get the cached version. But since the cached data is much older
there is a chance meanwhile there will be changes and we will serve
the wrong data.