Show inserted entry immediately
Don't wait for another AJAX get_recent_entries()
call before showing the just inserted entry.

Instead, return the inserted entry in the response
of the insert AJAX call.

The only trick is to not send latest_timestamp,
because otherwise the client won't query for
entries older than the inserted one, but there 
might be some between this moment and the previous
update.