Adjust request time to server time, instead of only to local
In case the local time is off from the server time, we may get wrong
data and also our content might not be cached.

To solve this after each successful server response we keep the current
server time and the current local time.

Then on the next request we use the server time as a base and just add
the different between the current local time and the local time from
the last response.

This way we limit our error to the error in time difference between two
requests, which is unlikely to be as big as a second and thus unlikely
to cause much trouble.