Don't include event entries if we don't need them
Event entries are those which replace some other
entry. When the replaced entry is in the query
result set we don't need the replacing entry,
because the replaced entry should be already
updated with its new value.

As a consequence of the filter implementation,
the query result is now keyed by the comment id.