Filter event entries as late as possible
We were filtering event entries every time we
called get(), but we use get() for every query,
so we never got the event entries.

Now we filter in the end of the queries. Since
we have only two, it's still not a big burden.

For the cases when we need all filtered entries
we introduce get_all() and to stop the confusion
of which one should we use outside of the class
we make get() private.