Commit entry modifications
Currently, we rely on the event entries to modify
the stale edited ones. But when we refresh this
is a lot harder and error-prone.

That's why we just commit the changes and when
event entries and its edited entry happen to be
in the same result set, we just skip the event
entry.