Shut the auto-update only on consecutive failures
Before, it we reached the max retries in total we would have shut down
the auto-update, which doesn't make much sense, becasue it can go up
and down and be working for hours and the number of failures can slowly
accumulate.

To convey this semantics better, we renamed all retries/failures
variables to include the word "consecutive".