Kses posts content closer to the place we use it
We were ksesing the contents right after we got it
from POST, which caused a post update getting empty
after uses to be deleted, instead of empty.

This isn't very DRY, but we need to extract the
CRUD functionality for entries anyway, we will
have this in central place then.