Use the Date header to get server-side date
Before, we were passing the current server timestamp in the JSON.
But when the JSON is cached and a client is a little bit behind, they
may get a stale timestamp, which is awful.

Instead, we will be getting the server timestamp via the Date header,
which is required by all servers with clocks. See
http://pretty-rfc.herokuapp.com/RFC2616#header.date