Flush the rewrite rules on rules update
Calling `add_rewrite_endpoint()` isn't enough to
add a rewrite rule, we need to flush the rules,
too. But we can't do this on every page load,
because it's a database write.

Instead, we keep the version of the liveblog
rewrite rules both in the code and in the database.
If they don't match this means there's an update
or we just installed the plugin and we need to
flush the rules to the database.

Fixes #11, props @danielbachhuber