Add correct endpoing to draft URLs with query args
Before we added /liveblog/ to a permalink if the
blog had pretty permalinks. But if the post was
a draft it's URL isn't pretty, it's ?p=XX-style.

So, instead of checking whether the blog used
pretty permalinks we check if it has ?p=. There
shouldn't be any false positives, because if a URL
has a ? and it doesn't mean a separator between
path and query args it should be escaped.

Fixes #17.