WPCOM_Liveblog_Entry::insert()
It's a static, because it cannot work on an instance.

Since all of the CRUD methods will need to insert a
comment, I added a private insert_comment() method,
too.

Some of the error reporting is crude. For example
we always send a server error, no matter what
broke. Differentiating will be fragile and it
doesn't matter so much.

The update/delete AJAX calls won't work, because
we removed the support until we implement them in
WPCOM_Liveblog_Entry.