Apply modifying entries to the queue
Imagine the following scenario:
 * An editor posts an entry
 * The user gets a nag with 1 new entry
 * The editor deletes the just-added entry
 * The entry is deleted, but since the queue isn't
touched, the user still sees the same message.

In this commit we apply all new modifying (update/delete)
entries to the queue. If we get a delete entry for an
entry in the queue, we remove it from the queue. If we
get an update entry, we just update the html of the
entry in the queue. Both of these trigger a re-render
of the fixed nag and it's always correct.