Show error callback status for non-HTTP errors
jQuery provides us a status argument for error
callbacks, which in case of HTTP errors is just
error, but can be something else if it's not an
HTTP error: abort, parse error, timeout, etc.

In case of missing or 200 or less HTTP code we can
assume that it's not an HTTP error, so we use
shorter message and just give the jQuery-provided
status.

Fixes #66.