Use human time diff instead of absolute timestamps
When viewing a liveblog the absolute time isn't
so important, we mostly care how old is an entry.

Here we introduced moment.js, which takes care of
that.

When a liveblog is archived, the users will see the
absolute timestamps. Since the difference will be
huge, it doesn't make so much sense to show the
difference in this case. Also, this behaviour comes
for free, because we don't show the relative
timestamps before we initialise the liveblog, which
never happens if it's archived.

Fixes #33.