Don't give the user to edit rendered content
We render each entry – substitute shortcodes, YouTube
videos, add paragraphs, etc.

Currently, when we edit it, the rendered contents are
offered for edit. This is not very user-friendly, because
some of the substitutions are very unwieldily.

That's why we are now passing the original content, so
that we can give it to the user for editing.