Switch the jQuery mess in publisher to Backbone
In our case this fixes the problem with cloning the
insert/edit form. Having it as a view makes creating
and removing forms a breeze.

We still need to hook it with the Entry model, but
it's already more readable now.

Also, we still need to make the preview working,
since jQuery UI tabs work only with divs with IDs,
not classes.