Fix for comment's output html and trash status
This fix should once and for all fix all the issues with the deleted entries.
The previous ajax html output was including the new comment id which upon
any delete action wasn't marking the first message with 'trash'. And on
edit actions wasn't setting the correct meta_value for the first comment_id.

Also fixed the timestamp returned by the edit action which always included
the current timestamp and no the first comment entry.