Enable WebSocket support only if server is using PHP >= 5.3.0
This commit checks the PHP version and load WebSocket support only if server is running PHP >= 5.3.0 (which is required by socket.io-php-emitter). If an older version is found, a message is added to the admin asking the user to update PHP or to remove LIVEBLOG_USE_SOCKETIO from wp-config.php.