Make update nag bar work when Socket.io support is enabled
The update nag bar was not being used when Socket.io support was enabled as liveblog.display_entry() was being called directly. This commit separates the logic to use the update nag bar from the logic that handles the AJAX response by creating a new method called liveblog.maybe_display_entries() and uses it when the Socket.io client receives a new message.