Use WebSocket to update the list of entries for private Liveblog posts as well
Before this commit WebSocket was used to update the list of entries only for public Liveblog posts. Now it works for private Liveblog posts as well.

To achieve this WordPress generates a unique key for each post based on the post ID and its status and share this key only with the clients who have permission to see the post. The clients forward this key to the Socket.io server. The Socket.io server use the key to add the clients to a room named after the post key. WordPress emits a new message with a Liveblog post entry only to the clients that are in the post room. This way only users who can see the post will receive the messages.