Add filter called 'liveblog_socketio_post_key' to allow a custom post key to be generated
When WebSocket is enabled, the plugin uses an unique key for each post based on its ID and status to control who can receive messages emitted by the Socket.io server. This should be enough for most cases but it has its limitations. Since the key for a given post is the same for all users and it doesn't change after a period of time, once a user receives the key it can have access to the messages emitted by the Socket.io server forever (assuming he saved the key somewhere) even if his account is removed or he loses access to the post.

This is not a problem for public Liveblog posts but it can be for private Liveblog posts with sensitive data. So the filter 'liveblog_socketio_post_key' was created to allow sites to implement their own criteria to generate the key. For example, a site could implement a random post key that can be manually invalidated by an editor whenever necessary.