Fixes oembed links which are not working properly when publishing from chrome/safari
contenteditable attribute in Chrome and Safari is wrapping new lines to div elements and thus links which otherwise would be standing on their own line are being wrapped in div and are not being picked up by autoembed feature of WordPress since it only takes into account links either standing on a line on their own, or wrapped in <p> element.

This commit is adding a new callback for liveblog_before_insert_entry filter which uses the same regular expression as core does for identifying link wrapped in <p> element (just is using div instead of p).

The preg_replace replaces the <div> element for <p> and core's autoembed feature then picks those links without issues.