Move is_robot function to javascript
The Liveblog is currently reyling on an existence of `vary_cache_on_function` function when modifying the behaviour for bots.

As not all environments taking advantage of page level caching have the posibility of varying page cache on user agent, the check should happen in browser.

This commit it moving the logic to inside the js/liveblog-lazyloader.js taking advantage of the same regex and disabling the lazyloading initialisation for robots + hiding the load more button.