Cast liveblog rewrite version before checks
The option is returned is a `string` and the class holds the version as an `int`. The strict check causes the version check to fail and we end up flushing rewrites every time `init` fires.

Before doing the check, cast the rewrite version option as an `int` to avoid the type mismatch.