Add state & progress indicators to an Entries Delete btn Re: #476
When a user on a slower network/server clicked delete, there was no 
visual feedback that their action was processing.

Adding this state allows ternary checks within the component to change 
the buttons text & add a 'active' CSS animation class.

This gives the user a indicator of their delete action is being 
preformed.