Add isPublishingNew state to reference new submissions progress Re: #476
When a user on a slower network/server clicked Publish New Entry, there 
was no visual feedback that their action was processing.

Adding this state by using pre-existing Redux actions, we can know when 
the Publishing transaction is occurring, allowing ternary checks within 
the components, which gives the availability to change the buttons text 
and classes.