Add isPublishing state to Entries to reference update progress Re: #476
When a user on a slower network/server clicked Publish Update, there was 
no visual feedback that their action was processing.

Adding this state to the entries by using pre-existing Redux actions, we 
can know when the Publishing transaction is occurring, allowing 
conditional checks within the components. This gives the option to 
change the buttons text and classes, which can indicate progress.