Add progress indicators to the publish btn on add/update Entry Re: #476
When a user on a slower network/server clicked Publish Update, there was 
no visual feedback that their action was processing.

By using the isPublishingNew state for the main 'add new entry' form, 
and using the isPublish state while editing the contents of individual 
entries, the component can now change the button text and buttons css 
class to indicate the action is in progress.