Remove entryEditClose function from edit entry editor. Re: #476
This function was used to close the editor immediately after clicking 
'Publish Update'. When a user on a slower network/server clicked this, 
the editor disappeared instantly, however it would take several seconds 
for the entries text to actually update - there was no visual feedback 
that their action was processing.

By not using this function the editor does not instantly get removed. 
Not being removed right away allows the editor component to check if the 
`store > user > entries` entry has a true isPublishing state, adding 
necessary indicators of progress.

The editor is already set to close by changing the isEditing state to 
false, on the UPDATE_ENTRY_SUCCESS Redux action in the user.js reducer.