add sanitization to dangerouslySetInnerHTML values. Re: #526
There are several spots in this plugin using Reacts 
`dangerouslySetInnerHTML` attribute to place content from the server. 
This can easily lead to cross-site scripting (XSS) attacks.

Before the HTML is dangerously placed, running it through DOMPurify 
https://github.com/cure53/DOMPurify helps to sanitize out malicious 
HTML.

iframe (an easy method for XSS), are allowed for the embedded media, 
however the domain used must be whitelisted - otherwise the iframes src 
is removed.