AMP: Fix too high refresh interval
The refresh interval can be set by the site owner to any value. On AMP liveblogs this is used for the `data-poll-interval` in the `amp-live-list` but that value has to be greater than 15. https://amp.dev/documentation/examples/components/amp-live-list/

This change ensures that AMP Liveblogs enforce a minimum value of 15 for the interval.